rafactor auth
