--- client_secret_example.json
@@ -1,3 +1,0 @@
-{
-    "Note": "Place your downloaded client_secret.jsonfrom google api here"
- }
